refactor: добавлена проверка на null releaseDate; добавлен общий интерфейс для классов сервиса, имеющих методы getAll и getById

+++ src/main/java/ru/yandex/practicum/filmorate/storage/genre/GenreDaoStorage.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface GenreDaoStorage {
     List<Genre> getAllGenre();
-    Genre getGenreById(Integer id);
+    Genre getGenreById(Long id);
 }
