+++ src/main/java/ru/yandex/practicum/filmorate/storage/rating/MpaDaoStorage.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface MpaDaoStorage {
     List<Mpa> getAllRating();
-    Mpa getRatingById(Integer id);
+    Mpa getRatingById(Long id);
 }
